reorganized nodemap by removing functions from useeffect. started reduxtoolkit setup. reorganized components into subfolders.
