+++ package.json
@@ -17,6 +17,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@reduxjs/toolkit": "^1.9.7",
     "@types/react": "^18.2.28",
     "@types/react-dom": "^18.2.13",
     "cors": "^2.8.5",
@@ -26,6 +27,7 @@
     "html-webpack-plugin": "^5.5.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-redux": "^8.1.3",
     "react-router-dom": "^6.17.0",
     "router-dom": "^2.2.11",
     "ts-loader": "^9.5.0",
