+++ renderer/src/components/AnnotationMenu/AnnotationMenu.tsx
@@ -1,0 +1,1 @@
+// Component to render the annotation menu
