+++ renderer/src/index.tsx
@@ -5,26 +5,30 @@ import {
   createBrowserRouter,
   RouterProvider,
 } from 'react-router-dom';
-import NodeMap from './components/NodeMap';
+import NodeMap from './components/NodeMap/NodeMap';
 import ViewCluster from './components/ViewCluster';
 
 
+// Configure routes using createBrowserRouter
+// Define paths and the components to render for each path
 const router = createBrowserRouter([
   {
-    path: "/",
-    element: <App />,
+    path: "/", // Root path
+    element: <App />, // Render App component at root path
   },
   {
-    path: "/viewlogin",
-    element: <NodeMap />,
+    path: "/viewlogin", // Path for login view
+    element: <NodeMap />, // Render NodeMap component
   },
   {
-    path: "/viewlogin",
-    element: <ViewCluster />,
+    path: "/viewlogin", // Duplicate path - might lead to issues
+    element: <ViewCluster />, // Render ViewCluster component
   },
 ]);
 
+// Create a React root attached to the DOM element with id 'root'
 const root = ReactDOM.createRoot(document.getElementById('root'));
+// Render the app within RouterProvider to enable routing
 root.render(
   <RouterProvider router={router} />
 );
