+++ renderer/src/styles/base.css
@@ -1,0 +1,7 @@
+body {
+    background-image: url('../assets/k.jpeg');
+    background-size: cover;
+    background-repeat: no-repeat;
+    background-attachment: fixed;
+    background-color: aquamarine;
+}
